Move cloaked knight to shared .h file
